+++ start/src/App.vue
@@ -1,9 +1,17 @@
 <script lang="ts" setup>
   import { ref } from 'vue'
 
-  const files1 = ref([])
+  const show1 = ref(true)
+
+  const listImg = [
+    'https://tianyuhao.cn/images/auto/1.jpg',
+    'https://tianyuhao.cn/images/auto/2.jpg',
+    'https://tianyuhao.cn/images/auto/3.jpg',
+    'https://tianyuhao.cn/images/auto/4.jpg'
+  ]
 </script>
 
 <template>
-  <f-up-load v-model:files="files1" />
+  <f-button type="primary" @click="show1 = true">打开</f-button>
+  <f-image-preview v-model:visible="show1" :img-list="listImg" />
 </template>